+++ openlogin/setup-2fa/index.test.ts
@@ -1,8 +1,39 @@
-import { expect } from "@playwright/test";
+import { expect, Page } from "@playwright/test";
 import { test } from "./index.lib";
 import { confirmEmail } from "../../utils";
-import { setup2FA } from "../../utils";
 import { useAutoCancelShareTransfer } from "../../utils/index";
+import * as fs from "fs";
+
+async function setup2FA(page: Page, flow: string) {
+  try {
+    await Promise.all([
+      page.waitForNavigation(/*{ url: 'https://app.openlogin.com/register#upgrading=true' }*/),
+      page.click('button:has-text("Enable 2FA")'),
+    ]);
+    await page.click(".v-input--selection-controls__ripple");
+    await page.click('button:has-text("Save current device")');
+    // await page.click('button:has-text("View advanced option")');
+    await page.click("text=View advanced option");
+    const [download] = await Promise.all([
+      page.waitForEvent("download"),
+      // page.click('button:has-text("Download my recovery phrase")'),
+      page.click("text=Download my recovery phrase"),
+    ]);
+    const downloadedFile = await download.path();
+    const backupPhrase = fs.readFileSync(downloadedFile, "utf8");
+    await page.click('button:has-text("Continue")');
+
+    await page.fill("textarea", backupPhrase);
+    await page.click('button:has-text("Verify")');
+    await Promise.all([
+      page.waitForNavigation(/*{ url: 'https://app.openlogin.com/wallet/home' }*/),
+      page.click('button:has-text("Done")'),
+    ]);
+    return true;
+  } catch {
+    return false;
+  }
+}
 
 test.describe("Setup 2FA", () => {
   test("Setup 2FA", async ({ browser, openloginURL, user }) => {
@@ -10,14 +41,13 @@ test.describe("Setup 2FA", () => {
     const page = await context.newPage();
     await page.goto(openloginURL);
     await page.click('button:has-text("Get Started")');
-    let testUser = user.email_2fa_login;
     let to = user.email_2fa_login;
     // Login with Passwordless
     const timestamp = Math.floor(Date.now() / 1000);
-    await page.fill('[placeholder="Email"]', testUser);
+    await page.fill('[placeholder="Email"]', to);
     await page.click('button:has-text("Continue with Email")');
     await page.waitForSelector("text=email has been sent");
-    expect(await page.isVisible(`text=${testUser}`)).toBeTruthy();
+    expect(await page.isVisible(`text=${to}`)).toBeTruthy();
     console.log(`user from test: ${to}`);
     // Confirm email
     test.fixme(
